feat: 시작 및 종료 날짜 String 타입으로 변경
